pyproject.toml
[project]
name = "celestializer"
version = "0.1.0"
description = "Processing and analysis tool for astronomical imaging"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "astrometry>=4.1.2",
    "astropy>=7.0.0",
    "astroquery>=0.4.8",
    "ipykernel>=6.29.5",
    "matplotlib>=3.9.3",
    "numpy>=2.1.3",
    "opencv-python>=4.10.0.84",
    "pandas>=2.2.3",
    "pillow>=11.0.0",
    "pydantic>=2.9.2",
    "rawkit>=0.6.0",
    "scikit-learn>=1.6.0",
    "tqdm>=4.67.1",
    "xxhash>=3.5.0",
]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["celestializer"]
package-dir = { "" = "src" }

[tool.uv]
dev-dependencies = ["celestializer"]

[tool.uv.sources]
celestializer = { workspace = true }
